Fix stale orderForm when auth state changed
This commit fixes an issue where if you changed the authentication state
in the orderForm, we would end up using the remote value returned by
Apollo's useQuery (which after identifying, is stale). This happened
because in the time this fix was implemented we had another approach to
managing both the remote and local state of the orderForm, which relied
more in the local state than in the remote one.

But now we have changed how we deal with it to always fetch the remote
orderForm if it is possible and wait for the query to resolve, instead
of "fallbacking" to the local value, and this guarantees that we will
always see the most up to date value of the orderForm, regardless of
these heuristics.